+++ src/platforms/shared/int.h
@@ -4,11 +4,11 @@
 #include "platforms/shared/int_base.h"
 
 namespace fl {
-    // On AVR: int is 16-bit, long is 32-bit — match stdint sizes manually
-    typedef int i16;
-    typedef unsigned int u16;
-    typedef long i32;
-    typedef unsigned long u32;
+    // Default platform (desktop/generic): assume short 16-bit, int 32-bit (uint32_t is unsigned int)
+    typedef short i16;
+    typedef unsigned short u16;
+    typedef int i32;             // 'int' is 32-bit on all desktop and most embedded targets
+    typedef unsigned int u32;
     typedef long long i64;
     typedef unsigned long long u64;
 }
